+++ backend/migrations/1739415955_deleted_user_preferences.go
@@ -1,0 +1,60 @@
+package migrations
+
+import (
+	"encoding/json"
+
+	"github.com/pocketbase/dbx"
+	"github.com/pocketbase/pocketbase/daos"
+	m "github.com/pocketbase/pocketbase/migrations"
+	"github.com/pocketbase/pocketbase/models"
+)
+
+func init() {
+	m.Register(func(db dbx.Builder) error {
+		dao := daos.New(db);
+
+		collection, err := dao.FindCollectionByNameOrId("{{collection_name}}")
+		if err != nil {
+			return err
+		}
+
+		return dao.DeleteCollection(collection)
+	}, func(db dbx.Builder) error {
+		jsonData := `{
+			"id": "{{collection_name}}",
+			"created": "2025-02-13 02:58:35.056Z",
+			"updated": "2025-02-13 02:58:35.056Z",
+			"name": "user_preferences",
+			"type": "base",
+			"system": false,
+			"schema": [
+				{
+					"system": false,
+					"id": "{{field_id}}",
+					"name": "findings_filters",
+					"type": "json",
+					"required": false,
+					"presentable": false,
+					"unique": false,
+					"options": {
+						"maxSize": 0
+					}
+				}
+			],
+			"indexes": [],
+			"listRule": "@request.auth.id = user.id",
+			"viewRule": "@request.auth.id = user.id",
+			"createRule": "@request.auth.id = user.id",
+			"updateRule": "@request.auth.id = user.id",
+			"deleteRule": "@request.auth.id = user.id",
+			"options": {}
+		}`
+
+		collection := &models.Collection{}
+		if err := json.Unmarshal([]byte(jsonData), &collection); err != nil {
+			return err
+		}
+
+		return daos.New(db).SaveCollection(collection)
+	})
+}